+++ tests/unit/test_config.py
@@ -327,13 +327,13 @@ def test_configure(monkeypatch, settings, environment, other_settings):
             pretend.call(".domain"),
             pretend.call(".i18n"),
             pretend.call(".db"),
+            pretend.call(".tasks"),
             pretend.call(".rate_limiting"),
             pretend.call(".static"),
             pretend.call(".policy"),
             pretend.call(".search"),
             pretend.call(".aws"),
             pretend.call(".gcloud"),
-            pretend.call(".tasks"),
             pretend.call(".sessions"),
             pretend.call(".cache.http"),
             pretend.call(".cache.origin"),
